call_tools/requirements.txt
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hydra-core==1.3.2
idna==3.4
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
omegaconf==2.3.0
orjson==3.9.5
packaging==23.1
pluggy==1.3.0
pocketsphinx==5.0.2
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydub==0.25.1
pytest==7.4.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.16.2
